requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.2.1
motor==2.5.1
pymongo==3.12.3
bcrypt==4.1.3
PyJWT==2.10.1
python-jose==3.5.0
email-validator==2.3.0
dnspython==2.7.0